django_backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
google-auth==2.23.4
firebase-admin==6.2.0
python-dotenv==1.0.0
google-generativeai==0.3.1
